Remove Prefect
Since Prefect keeps ending up cranky, and isn't giving us a whole lot extra functionality while adding complexity, I've ripped it out and replaced it with Celery Beat for periodic task scheduling. I've additionally added Celery Flower for monitoring purposes.

Additionally some tweaks to the Docker configs to try to deal with some run-away CPU usage and Docker locking up entirely.
