+++ Readme.md
@@ -116,8 +116,6 @@ DJANGO_DEBUG=True
 DJANGO_MANAGEPY_MIGRATE=off
 SENTRY_DSN=some_long_string_from_sentry
 CELERY_BROKER_URL=rediss://cache:6379/1
-PREFECT__CLOUD__API_KEY=api-key
-PREFECT_PROJECT_NAME=project_that_the_prefect_flows_should_be_registered_to
 ```
 
 An additional environment variable is currently configured in `docker-compose.yaml` as it is unlikely to need to be changed.
