+++ app/deployments/tasks/queue.py
@@ -1,0 +1,24 @@
+def task_queued(task_name: str, task_args: list, task_kwargs: dict) -> bool:
+    """Returns true if the task is already scheduled"""
+    from buoy_barn.celery import app
+    from celery.app.control import Control
+
+    control = Control(app)
+    inspect = control.inspect()
+
+    for worker_tasks in inspect.active().values():
+        for task in worker_tasks:
+            if task["name"] == task_name and task["args"] == task_args:
+                return True
+
+    for worker_tasks in inspect.reserved().values():
+        for task in worker_tasks:
+            if task["name"] == task_name and task["args"] == task_args:
+                return True
+
+    for worker_tasks in inspect.scheduled().values():
+        for task in worker_tasks:
+            if task["name"] == task_name and task["args"] == task_args:
+                return True
+
+    return False
