+++ app/pyproject.toml
@@ -26,9 +26,9 @@ vcrpy = "^5.1"
 whitenoise = "^6.6"
 xarray = "^2023.12.0"
 pandas = "^1.5.3"
-prefect = "^1.4.0"
 django-health-check = "^3.17.0"
 pystac = "^1.9.0"
+flower = "^2.0.1"
 
 [tool.poetry.dev-dependencies]
 ipython = "^8.19"
