+++ k8s/base/celery-flower-service.yaml
@@ -1,0 +1,13 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: celery-flower
+spec:
+  selector:
+    service: celery
+    tier: backend
+    role: flower
+  ports:
+    - port: 5555
+      targetPort: 5555
+      protocol: TCP
