+++ k8s/base/config.env
@@ -1,3 +1,2 @@
 REDIS_CACHE=redis://buoy-barn-cache:6379/
 CELERY_BROKER_URL=redis://buoy-barn-cache:6379/1
-PREFECT_PROJECT_NAME=buoy-barn-prod
